add linter tests
