+++ openbb_platform/core/tests/app/static/test_package_builder.py
@@ -3,21 +3,23 @@
 
 from dataclasses import dataclass
 from inspect import _empty
-from unittest.mock import patch
+from pathlib import Path
+from unittest.mock import PropertyMock, mock_open, patch
 
 import pandas
 import pytest
+from importlib_metadata import EntryPoint, EntryPoints
 from openbb_core.app.static.package_builder import (
     ClassDefinition,
     DocstringGenerator,
     ImportDefinition,
-    Linters,
     MethodDefinition,
     ModuleBuilder,
     PackageBuilder,
     Parameter,
     PathHandler,
 )
+from openbb_core.env import Env
 from pydantic import Field
 from typing_extensions import Annotated
 
@@ -442,37 +444,6 @@ def test_build_module_class(path_handler):
 
 
 @pytest.fixture(scope="module")
-def linters(tmp_package_dir):
-    """Return linters."""
-    return Linters(tmp_package_dir)
-
-
-def test_linters_init(linters):
-    """Test linters init."""
-    assert linters
-
-
-def test_print_separator(linters):
-    """Test print separator."""
-    linters.print_separator(symbol="AAPL")
-
-
-def test_run(linters):
-    """Test run."""
-    linters.run(linter="ruff")
-
-
-def test_ruff(linters):
-    """Test ruff."""
-    linters.ruff()
-
-
-def test_black(linters):
-    """Test black."""
-    linters.black()
-
-
-@pytest.fixture(scope="module")
 def docstring_generator():
     """Return package builder."""
     return DocstringGenerator()
@@ -541,3 +512,129 @@ def test_generate(docstring_generator):
     assert doc
     assert "Parameters" in doc
     assert "Returns" in doc
+
+
+def test_read_extension_map(package_builder, tmp_package_dir):
+    """Test read extension map."""
+
+    PATH = "openbb_core.app.static.package_builder."
+    open_mock = mock_open()
+    with patch(PATH + "open", open_mock), patch(PATH + "load") as mock_load:
+        package_builder._read_extension_map(tmp_package_dir)
+        open_mock.assert_called_once_with(Path(tmp_package_dir, "extension_map.json"))
+        mock_load.assert_called_once()
+
+
+@pytest.mark.parametrize(
+    "ext_built, ext_installed, ext_inst_version, expected_add, expected_remove",
+    [
+        (
+            {
+                "openbb_core_extension": [
+                    "ext_1@0.0.0",
+                    "ext_2@0.0.0",
+                ],
+                "openbb_provider_extension": [
+                    "prov_1@0.0.0",
+                    "prov_2@1.1.1",
+                ],
+            },
+            EntryPoints(
+                (
+                    EntryPoint(
+                        name="ext_2", value="...", group="openbb_core_extension"
+                    ),
+                    EntryPoint(
+                        name="prov_2", value="...", group="openbb_provider_extension"
+                    ),
+                )
+            ),
+            "0.0.0",
+            {"prov_2@0.0.0"},
+            {"ext_1@0.0.0", "prov_1@0.0.0", "prov_2@1.1.1"},
+        ),
+        (
+            {
+                "openbb_core_extension": ["ext_1@9.9.9"],
+                "openbb_provider_extension": ["prov_2@0.0.0"],
+            },
+            EntryPoints(
+                (
+                    EntryPoint(
+                        name="ext_2", value="...", group="openbb_core_extension"
+                    ),
+                    EntryPoint(
+                        name="prov_1", value="...", group="openbb_provider_extension"
+                    ),
+                )
+            ),
+            "5.5.5",
+            {"ext_2@5.5.5", "prov_1@5.5.5"},
+            {"ext_1@9.9.9", "prov_2@0.0.0"},
+        ),
+    ],
+)
+def test_package_diff(
+    package_builder,
+    tmp_package_dir,
+    ext_built,
+    ext_installed,
+    ext_inst_version,
+    expected_add,
+    expected_remove,
+):
+    """Test package differences."""
+
+    def mock_entry_points(group):
+        return ext_installed.select(**{"group": group})
+
+    PATH = "openbb_core.app.static.package_builder."
+    with patch.object(
+        PackageBuilder, "_read_extension_map"
+    ) as mock_read_extension_map, patch(
+        PATH + "entry_points", mock_entry_points
+    ), patch.object(
+        EntryPoint, "dist", new_callable=PropertyMock
+    ) as mock_obj:
+
+        class MockPathDistribution:
+            version = ext_inst_version
+
+        mock_obj.return_value = MockPathDistribution()
+        mock_read_extension_map.return_value = ext_built
+
+        add, remove = package_builder._diff(tmp_package_dir)
+
+        # We add whatever is not built, but is installed
+        assert add == expected_add
+        # We remove whatever is built, but is not installed
+        assert remove == expected_remove
+
+
+@pytest.mark.parametrize(
+    "add, remove, openbb_auto_build",
+    [
+        (set(), set(), True),
+        ({"this"}, set(), True),
+        (set(), {"that"}, True),
+        ({"this"}, {"that"}, True),
+        ({"this"}, {"that"}, False),
+    ],
+)
+def test_auto_build(package_builder, tmp_package_dir, add, remove, openbb_auto_build):
+    """Test auto build."""
+
+    with patch.object(PackageBuilder, "_diff") as mock_package_diff, patch.object(
+        PackageBuilder, "build"
+    ) as mock_build, patch.object(Env, "AUTO_BUILD", openbb_auto_build):
+        mock_package_diff.return_value = add, remove
+
+        package_builder.auto_build(tmp_package_dir)
+
+    if openbb_auto_build:
+        mock_package_diff.assert_called_once_with(Path(tmp_package_dir, "package"))
+        if add or remove:
+            mock_build.assert_called_once()
+    else:
+        mock_package_diff.assert_not_called()
+        mock_build.assert_not_called()
